refactor(release-parsing): lift parsers out of modules/media
The four pure release-name parsers — quality, language, season/episode,
attributes — lived under \`modules/media/\` but had no domain coupling
to media: they take a string and return structured data. Auto-grab,
disk import, subtitle scorer, file rescan, and the upcoming torrent
auto-matcher all consume them — none of those are "media domain"
clients.

Moves them to a shared \`common/release-parsing/\` package with a
single barrel export. No behaviour change; mechanical rename + import
update across the 7 call sites:

  modules/media/release-quality.parser.ts      → common/release-parsing/quality.parser.ts
  modules/media/release-language.parser.ts     → common/release-parsing/language.parser.ts
  modules/media/release-episode.parser.ts      → common/release-parsing/season-episode.parser.ts
  modules/media/release-attributes.parser.ts   → common/release-parsing/attributes.parser.ts
  modules/media/release-episode.parser.spec.ts → common/release-parsing/season-episode.parser.spec.ts

\`release-rejection.helper.ts\` stays in \`modules/media/\` — it mixes
scoring against profiles (impure) with parsing and that coupling is
real.
