+++ backend/src/common/release-parsing/season-episode.parser.spec.ts
@@ -1,4 +1,4 @@
-import { parseSeasonEpisode } from './release-episode.parser';
+import { parseSeasonEpisode } from './season-episode.parser';
 
 describe('parseSeasonEpisode', () => {
   it('parses SxxExx', () => {
